TANGO-2130 : Updated and refactored tests
