--- hermes-consumers/src/test/resources/2130/int.avsc
@@ -1,11 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "Primitives",
-  "type": "record",
-  "fields": [
-    {
-      "name": "int_field",
-      "type": "int"
-    }
-  ]
-}
