--- hermes-consumers/src/test/resources/2130/int.proto
@@ -1,5 +1,0 @@
- syntax="proto3";
-
-message Primitives {
-  required int32 int_field;
-}
