--- hermes-consumers/src/test/resources/schemas/enum.avsc
@@ -1,18 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "Enum",
-  "type": "record",
-  "fields": [
-    {
-      "name": "enum_field",
-      "type": {
-        "name": "EnumTypeName",
-        "type": "enum",
-        "symbols": [
-          "S1",
-          "S2"
-        ]
-      }
-    }
-  ]
-}
