--- hermes-consumers/src/test/resources/schemas/evolution/mode-changed-from-nullable-to-required-2.avsc
@@ -1,12 +1,0 @@
-{
-  "name": "test",
-  "doc": "Doc",
-  "type": "record",
-  "fields": [
-    {
-      "name": "test",
-      "type": "string",
-      "doc": "test"
-    }
-  ]
-}
