--- hermes-consumers/src/test/resources/schemas/evolution/newly-added-nullable-record-with-required-fields-2.avsc
@@ -1,29 +1,0 @@
-{
-  "name": "test",
-  "doc": "Doc",
-  "type": "record",
-  "fields": [
-    {
-      "name": "test",
-      "type":["null", "string" ],
-      "doc": "test",
-      "default": null
-    },
-    {
-      "name": "rec",
-      "doc": "record",
-      "type": ["null", {
-        "type": "record",
-        "name": "Test2",
-        "fields": [
-          {
-            "name": "test2",
-            "doc": "test2",
-            "type": "string"
-          }
-        ]
-      }],
-      "default": null
-    }
-  ]
-}
