--- hermes-consumers/src/test/resources/schemas/evolution/newly-added-record-with-nested-required-fields-2.avsc
@@ -1,39 +1,0 @@
-{
-  "name": "test",
-  "doc": "Doc",
-  "type": "record",
-  "fields": [
-    {
-      "name": "test",
-      "type":["null", "string" ],
-      "doc": "test",
-      "default": null
-    },
-    {
-      "name": "rec",
-      "doc": "record",
-      "type": {
-        "type": "record",
-        "name": "Test2",
-        "fields": [
-          {
-            "name": "test2",
-            "doc": "test2",
-            "type": {
-              "name": "Nested",
-              "type": "record",
-              "doc": "test",
-              "fields": [
-                {
-                  "type": "string",
-                  "name": "nestedField",
-                  "doc": "nested"
-                }
-              ]
-            }
-          }
-        ]
-      }
-    }
-  ]
-}
