--- hermes-consumers/src/test/resources/schemas/evolution/newly-added-record-with-required-fields-1.avsc
@@ -1,13 +1,0 @@
-{
-  "name": "test",
-  "doc": "Doc",
-  "type": "record",
-  "fields": [
-    {
-      "name": "test",
-      "type":["null", "string" ],
-      "doc": "test",
-      "default": null
-    }
-  ]
-}
