--- hermes-consumers/src/test/resources/schemas/logical-types/bytesDecimal.avsc
@@ -1,16 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "bytesDecimal",
-      "type": {
-        "type": "bytes",
-        "logicalType": "decimal",
-        "precision": 14,
-        "scale": 2
-      }
-    }
-  ]
-}
