--- hermes-consumers/src/test/resources/schemas/logical-types/bytesDecimalJavaClass.avsc
@@ -1,17 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "bytesDecimal",
-      "type": {
-        "type": "bytes",
-        "logicalType": "decimal",
-        "java-class": "java.math.BigDecimal",
-        "precision": 14,
-        "scale": 2
-      }
-    }
-  ]
-}
