--- hermes-consumers/src/test/resources/schemas/logical-types/fixedDecimal.avsc
@@ -1,18 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "fixedDecimal",
-      "type": {
-        "name": "Fixed",
-        "type": "fixed",
-        "logicalType": "decimal",
-        "precision": 4,
-        "scale": 2,
-        "size": 8
-      }
-    }
-  ]
-}
