--- hermes-consumers/src/test/resources/schemas/logical-types/fixedDecimalJavaClass.avsc
@@ -1,19 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "fixedDecimal",
-      "type": {
-        "name": "Fixed",
-        "type": "fixed",
-        "logicalType": "decimal",
-        "java-class": "java.math.BigDecimal",
-        "precision": 4,
-        "scale": 2,
-        "size": 8
-      }
-    }
-  ]
-}
