--- hermes-consumers/src/test/resources/schemas/logical-types/fixedDuration.avsc
@@ -1,16 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "fixedDuration",
-      "type": {
-        "name": "fixedDuration",
-        "type": "fixed",
-        "size": 12,
-        "logicalType": "duration"
-      }
-    }
-  ]
-}
