--- hermes-consumers/src/test/resources/schemas/logical-types/intDate.avsc
@@ -1,14 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "intDate",
-      "type": {
-        "type": "int",
-        "logicalType": "date"
-      }
-    }
-  ]
-}
