--- hermes-consumers/src/test/resources/schemas/logical-types/intTimeMillis.avsc
@@ -1,14 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "intTimeMillis",
-      "type": {
-        "type": "int",
-        "logicalType": "time-millis"
-      }
-    }
-  ]
-}
