--- hermes-consumers/src/test/resources/schemas/logical-types/intTimeMillisJavaClass.avsc
@@ -1,15 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "intTimeMillis",
-      "type": {
-        "type": "int",
-        "logicalType": "time-millis",
-        "java-class": "java.time.LocalTime"
-      }
-    }
-  ]
-}
