--- hermes-consumers/src/test/resources/schemas/logical-types/longTimeMicros.avsc
@@ -1,14 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "longTimeMicros",
-      "type": {
-        "type": "long",
-        "logicalType": "time-micros"
-      }
-    }
-  ]
-}
