--- hermes-consumers/src/test/resources/schemas/logical-types/longTimeMicrosJavaClass.avsc
@@ -1,15 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "longTimeMicros",
-      "type": {
-        "type": "long",
-        "logicalType": "time-micros",
-        "java-class": "java.time.LocalTime"
-      }
-    }
-  ]
-}
