--- hermes-consumers/src/test/resources/schemas/logical-types/longTimestampMicros.avsc
@@ -1,14 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "longTimestampMicros",
-      "type": {
-        "type": "long",
-        "logicalType": "timestamp-micros"
-      }
-    }
-  ]
-}
