--- hermes-consumers/src/test/resources/schemas/logical-types/longTimestampMicrosJavaClass.avsc
@@ -1,15 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "longTimestampMicros",
-      "type": {
-        "type": "long",
-        "logicalType": "timestamp-micros",
-        "java-class": "java.time.Instant"
-      }
-    }
-  ]
-}
