--- hermes-consumers/src/test/resources/schemas/logical-types/longTimestampMillis.avsc
@@ -1,14 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "longTimestampMillis",
-      "type": {
-        "type": "long",
-        "logicalType": "timestamp-millis"
-      }
-    }
-  ]
-}
