--- hermes-consumers/src/test/resources/schemas/logical-types/longTimestampMillisJavaClass.avsc
@@ -1,15 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "longTimestampMillis",
-      "type": {
-        "type": "long",
-        "logicalType": "timestamp-millis",
-        "java-class": "java.time.Instant"
-      }
-    }
-  ]
-}
