--- hermes-consumers/src/test/resources/schemas/logical-types/stringUuid.avsc
@@ -1,14 +1,0 @@
-{
-  "type": "record",
-  "name": "LogicalTypes",
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "fields": [
-    {
-      "name": "stringUuid",
-      "type": {
-        "type": "string",
-        "logicalType": "uuid"
-      }
-    }
-  ]
-}
