--- hermes-consumers/src/test/resources/schemas/map-with-values-map.avsc
@@ -1,17 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "MapNested",
-  "type": "record",
-  "fields": [
-    {
-      "name": "map_field",
-      "type": {
-        "type": "map",
-        "values": {
-          "type": "map",
-          "values": "string"
-        }
-      }
-    }
-  ]
-}
