--- hermes-consumers/src/test/resources/schemas/nested.avsc
@@ -1,36 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "Nested",
-  "type": "record",
-  "fields": [
-    {
-      "name": "level1Record",
-      "type": ["null", {
-        "type": "record",
-        "name": "Level3Record",
-        "fields": [{
-          "name": "level4Array",
-          "type": {
-            "type": "array",
-            "items": {
-              "type": "map",
-              "values": {
-                "type": "record",
-                "name": "Level5Record",
-                "fields": [
-                  {
-                    "name": "level6String",
-                    "default": "null",
-                    "type": [
-                      "null", "string"
-                    ]
-                  }
-                ]
-              }
-            }
-          }
-        }]
-      }]
-    }
-  ]
-}
