--- hermes-consumers/src/test/resources/schemas/null.avsc
@@ -1,11 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "Null",
-  "type": "record",
-  "fields": [
-    {
-      "name": "null_field",
-      "type": "null"
-    }
-  ]
-}
