--- hermes-consumers/src/test/resources/schemas/record-primitives/empty-inner-record.avsc
@@ -1,19 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "RecordPrimitives",
-  "type": "record",
-  "fields": [
-    {
-      "type": "string",
-      "name": "field"
-    },
-    {
-      "name": "test2",
-      "type": {
-        "name": "test2",
-        "type": "record",
-        "fields": []
-      }
-    }
-  ]
-}
