--- hermes-consumers/src/test/resources/schemas/record-primitives/empty-schema.avsc
@@ -1,7 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "RecordPrimitives",
-  "type": "record",
-  "fields": [
-  ]
-}
