--- hermes-consumers/src/test/resources/schemas/union/union-null-at-second-position.avsc
@@ -1,11 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "Enum",
-  "type": "record",
-  "fields": [
-    {
-      "name": "union_field",
-      "type": ["int", "null"]
-    }
-  ]
-}
