--- hermes-consumers/src/test/resources/schemas/union/union-without-null.avsc
@@ -1,11 +1,0 @@
-{
-  "namespace": "pl.allegro.bigdata.avro.gcpavrotools",
-  "name": "Enum",
-  "type": "record",
-  "fields": [
-    {
-      "name": "enum_field",
-      "type": ["int", "long"]
-    }
-  ]
-}
